requirement.txt
Django==4.0.5
django-cronjobs==0.2.3
django-crontab==0.7.1
djangorestframework==3.14.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
djangorestframework-simplejwt
requests
psycopg2-binary
drf-api-logger
whitenoise